pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "1.0.0"
description = "Jarvis 1.0 — training stack, memory, and autonomous agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
training = [
    "torch>=2.2",
    "transformers>=4.40",
    "deepspeed>=0.14",
    "datasets>=2.18",
    "tensorboard>=2.16",
]
vectors = [
    "sentence-transformers>=2.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ruff]
line-length = 100
target-version = "py311"
